uefi: Move various types to the `uefi::runtime` module
The types are publicly re-exported from `uefi::table::runtime`, so there should
be no breaking change to the public API.
